requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-js-asset==2.0.0
django-mptt==0.14.0
Faker==18.3.4
idna==3.4
pycodestyle==2.10.0
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
